Add storage proviioner in spec and seperated affinity in different file
